# doc/openapi.yml
openapi: 3.0.3

info:
  title: Madagascar Voting System API
  description: API for a mobile electronic voting system in Madagascar with biometric authentication and JWT
  version: 1.0.0

servers:
  - url: https://api.vote-madagascar.com

security:
  - BearerAuth: []

tags:
  -

paths:
  /citizens:
    post:
      operationId: createCitizen
      tags:
        - Citizen
      description: Create a citizen
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/CreateCitizen"
      responses:
        '201':
          description: Citizen(s) created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Citizen"

    get:
      operationId: getCitizens
      tags:
        - Citizen
      description: Get all citizens
      parameters:
        - in: query
          name: gid
          required: false
          schema:
            type: string
        - in: query
          name: id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Citizen"

    put:
      operationId: updateCitizens
      tags:
        - Citizen
      description: Update citizen(s) Info
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UpdateCitizen"
      responses:
        '200':
          description: Citizen(s) updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Citizen"

  /elections:
    post:
      operationId: createElections
      tags:
        - Election
      description: Create Election(s)
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/CreateElection"
      responses:
        '201':
          description: Election(s) created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Election"

  /elections/{electionId}/result:
    get:
      operationId: getElectionResult
      tags:
        - Election
      description: Get Election Result
      parameters:
        - in: path
          name: electionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionResult"

  /elections/{electionId}/blind-signature/public-key:
    get:
      operationId: getElectionBlindSignaturePublicKey
      tags:
        - Election
      description: Get Election Blind Signature Public Key
      parameters:
        - in: path
          name: electionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlindSignaturePublicKey"

  /elections/{electionId}/blind-signature/sign:
    post:
      operationId: signBlindedMessage
      tags:
        - Election
      description: Sign a blinded message for an election
      parameters:
        - in: path
          name: electionId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlindSignatureRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlindSignatureResponse"

  /votes:
    post:
      operationId: createVotes
      tags:
        - Vote
      description: Create Vote(s)
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/CreateVote"
      responses:
        '201':
          description: Vote(s) created

  /auth/login:
    post:
      operationId: login
      tags:
        - Auth
      description: Authenticate a citizen and get a JWT
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/register:
    post:
      operationId: register
      tags:
        - Auth
      description: Register a new citizen
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCitizen"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Citizen"

  /auth/me:
    get:
      operationId: whoami
      tags:
        - Auth
      description: Get current authenticated citizen
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Citizen"

components:
  schemas:
    LoginRequest:
      type: object
      properties:
        gid:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string

    ElectionResult:
      type: object
      properties:
        electionId:
          type: string
        totalVote:
          type: integer
        candidateResults:
          type: array
          items:
            $ref: "#/components/schemas/ElectionCandidateResult"

    ElectionCandidateResult:
      type: object
      properties:
        candidateGid:
          type: string
        voteAmount:
          type: integer
    CreateVote:
      type: object
      properties:
        electionId:
          type: string
        candidateId:
          type: string
        message:
          type: string
        signature:
          type: string

    BlindSignaturePublicKey:
      type: object
      properties:
        modulus:
          type: string
        publicExponent:
          type: string

    BlindSignatureRequest:
      type: object
      properties:
        blindedMessage:
          type: string

    BlindSignatureResponse:
      type: object
      properties:
        blindedSignature:
          type: string

    Election:
      allOf:
        - $ref: "#/components/schemas/CreateElection"
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426655440000
        createdAt:
          type: string
          format: date-time

    CreateElection:
      type: object
      properties:
        title:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/ElectionCandidate"

    ElectionCandidate:
      type: object
      properties:
        gid:
          type: string
          description: Gouvernement ID
        description:
          type: string

    UpdateCitizen:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426655440000
        firstName:
          type: string
        lastName:
          type: string

    Citizen:
      allOf:
        - $ref: "#/components/schemas/CreateCitizen"
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426655440000
        role:
          $ref: "#/components/schemas/Role"

    CreateCitizen:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        gid:
          type: string
          description: Gouvernement ID
        password:
          type: string

    Role:
      type: string
      enum:
        - ADMIN
        - USER

    BiometricData:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/BiometricDataType"
        embedding:
          type: array
          items:
            type: number
            format: float

    BiometricDataType:
      type: string
      enum:
        - FACE_ID

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT